+++ internal/next/websvc/settings_test.go
@@ -49,6 +49,9 @@ func TestService_HandleGetSettingsAll(t *testing.T) {
 	}
 
 	svc, err := websvc.New(&websvc.Config{
+		Pprof: &websvc.PprofConfig{
+			Enabled: false,
+		},
 		TLS: &tls.Config{
 			Certificates: []tls.Certificate{{}},
 		},